Creacion de Carpeta funciones.py

+++ casting.py
@@ -3,6 +3,6 @@ num1 = int(input())
 print("ingresa el segundo punto")
 num2 = int(input())
 print("ingresa el tercer punto")
-num3 = float(input())
+num3 = int(input())
 resultado = num1 + num2 + num3
 print("El resultado es:" + str(resultado))
